+++ src/integrationTest/java/com/mongodb/hibernate/BasicInsertionTests.java
@@ -16,6 +16,7 @@
 
 package com.mongodb.hibernate;
 
+import static com.mongodb.hibernate.internal.MongoAssertions.assertNotNull;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hibernate.cfg.JdbcSettings.JAKARTA_JDBC_URL;
 
@@ -120,7 +121,7 @@ class BasicInsertionTests {
                 .applyConnectionString(connectionString)
                 .build())) {
             var collection = mongoClient
-                    .getDatabase(MongoAssertions.assertNotNull(connectionString.getDatabase()))
+                    .getDatabase(assertNotNull(connectionString.getDatabase()))
                     .getCollection("books", BsonDocument.class);
             collectionConsumer.accept(collection);
         }
