+++ src/main/java/com/mongodb/hibernate/internal/translate/mongoast/AstLiteralValue.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.hibernate.internal.mongoast;
+package com.mongodb.hibernate.internal.translate.mongoast;
 
 import org.bson.BsonValue;
 import org.bson.BsonWriter;
