+++ src/main/java/com/mongodb/hibernate/internal/translate/mongoast/AstPlaceholder.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2024-present MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.mongodb.hibernate.internal.translate.mongoast;
+
+import org.bson.BsonWriter;
+
+/**
+ * Represents an MQL parameter placeholder, whose values are provided to {@link java.sql.PreparedStatement}'s various
+ * setter methods together with their position indexes.
+ *
+ * <p>Note that MQL has no SQL parameter placeholder (JDBC uses {@code ?} as placeholder marker) counterpart; currently
+ * the {@linkplain BsonWriter#writeUndefined() BSON Undefined} value is chosen.
+ *
+ * <p>This class is not part of the public API and may be removed or changed at any time
+ */
+public final class AstPlaceholder implements AstValue {
+
+    public static AstPlaceholder INSTANCE = new AstPlaceholder();
+
+    private AstPlaceholder() {}
+
+    @Override
+    public void render(BsonWriter writer) {
+        writer.writeUndefined();
+    }
+}
