+++ src/main/java/com/mongodb/hibernate/internal/translate/package-info.java
@@ -14,23 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.hibernate.internal.mongoast;
+@NullMarked
+package com.mongodb.hibernate.internal.translate;
 
-import org.bson.BsonWriter;
-
-/**
- * Represents some Bson field with name and value, which is usually rendered with other {@link AstElement}s to compose a
- * {@link org.bson.BsonDocument}.
- *
- * <p>This class is not part of the public API and may be removed or changed at any time
- *
- * @param name field name
- * @param value field value
- */
-public record AstElement(String name, AstValue value) implements AstNode {
-    @Override
-    public void render(BsonWriter writer) {
-        writer.writeName(name);
-        value.render(writer);
-    }
-}
+import org.jspecify.annotations.NullMarked;
