+++ src/test/java/com/mongodb/hibernate/internal/translate/AstVisitorValueHolderTests.java
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package com.mongodb.hibernate.translate;
+package com.mongodb.hibernate.internal.translate;
 
-import static com.mongodb.hibernate.translate.AstVisitorValueDescriptor.COLLECTION_MUTATION;
-import static com.mongodb.hibernate.translate.AstVisitorValueDescriptor.FIELD_VALUE;
+import static com.mongodb.hibernate.internal.translate.AstVisitorValueDescriptor.COLLECTION_MUTATION;
+import static com.mongodb.hibernate.internal.translate.AstVisitorValueDescriptor.FIELD_VALUE;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.mongodb.hibernate.internal.mongoast.AstElement;
-import com.mongodb.hibernate.internal.mongoast.AstLiteralValue;
-import com.mongodb.hibernate.internal.mongoast.AstPlaceholder;
-import com.mongodb.hibernate.internal.mongoast.command.AstInsertCommand;
+import com.mongodb.hibernate.internal.translate.mongoast.AstElement;
+import com.mongodb.hibernate.internal.translate.mongoast.AstLiteralValue;
+import com.mongodb.hibernate.internal.translate.mongoast.AstPlaceholder;
+import com.mongodb.hibernate.internal.translate.mongoast.command.AstInsertCommand;
 import java.util.List;
 import org.bson.BsonString;
 import org.junit.jupiter.api.BeforeEach;
